CtrlCore: HotKey implementation fixed to workaround MSC7.1 compiler bug (thanks glx117)
git-svn-id: svn://ultimatepp.org/upp/trunk@1592 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlCore/Win32Wnd.cpp
@@ -710,7 +710,7 @@ void Ctrl::sProcessMSG(MSG& msg)
 {
 	if (msg.message == WM_HOTKEY) {
 		if(msg.wParam >= 0 && (int)msg.wParam < Ctrl::hotkey.GetCount())
-			Ctrl::hotkey[msg.wParam]();
+			Ctrl::hotkey[(int)msg.wParam]();
 		return;
 	}
 	if(msg.message != WM_SYSKEYDOWN && msg.message != WM_SYSKEYUP
